refactored settings_data::LayerSpecifier.index_in_gamepad to use\n usize instead of u32

+++ src-tauri/src/settings_data.rs
@@ -257,7 +257,7 @@ pub enum SwitchOnClickReaction {
 #[derive(Debug, Clone, PartialEq)]
 pub struct LayerSpecifier {
     pub id: String,
-    pub index_in_gamepad: Option<u32>,
+    pub index_in_gamepad: Option<usize>,
 }
 
 impl LayerSpecifier {
